+++ script/build.ts
@@ -62,18 +62,28 @@ async function buildAll() {
   ];
   const externals = allDeps.filter((dep) => !allowlist.includes(dep));
 
-  await esbuild({
-    entryPoints: ["server/index.ts"],
-    platform: "node",
+  const esbuildOpts = {
+    platform: "node" as const,
     bundle: true,
-    format: "cjs",
-    outfile: "dist/index.cjs",
+    format: "cjs" as const,
     define: {
       "process.env.NODE_ENV": '"production"',
     },
     minify: true,
     external: externals,
-    logLevel: "info",
+    logLevel: "info" as const,
+  };
+
+  await esbuild({
+    entryPoints: ["server/index.ts"],
+    outfile: "dist/index.cjs",
+    ...esbuildOpts,
+  });
+
+  await esbuild({
+    entryPoints: ["server/worker.ts"],
+    outfile: "dist/worker.cjs",
+    ...esbuildOpts,
   });
 }